requirements.txt
beautifulsoup4==4.8.2
discord.py==1.2.5
lxml==4.4.2
requests==2.22.0